feat: add typeorm in project and config database connection
