+++ package.json
@@ -17,8 +17,11 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
+    "pg": "^8.11.2",
+    "reflect-metadata": "^0.1.13",
     "rimraf": "^5.0.1",
     "swagger-ui-express": "^5.0.0",
+    "typeorm": "^0.3.17",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
