+++ src/shared/http/server.ts
@@ -1,32 +1,10 @@
 import 'dotenv/config';
-import express, { NextFunction, Request, Response } from 'express';
-import swaggerUi from 'swagger-ui-express';
-import cors from 'cors';
-import { routes } from './routes';
-import { AppError } from '../errors/AppError';
-import swaggerFile from '../../swagger.json';
+import 'reflect-metadata';
+import { app } from './app';
+import { dataSource } from '../typeorm/index';
 
-const app = express();
-app.use(cors());
-app.use(express.json());
-app.use('/api/docs', swaggerUi.serve, swaggerUi.setup(swaggerFile));
-
-app.use(routes);
-
-app.use(
-  (error: Error, request: Request, response: Response, next: NextFunction) => {
-    if (error instanceof AppError) {
-      return response
-        .status(error.statusCode)
-        .json({ status: 'error', message: error.message });
-    }
-    console.log(error);
-    return response
-      .status(500)
-      .json({ status: 'error', message: 'Internal server error' });
-  },
-);
-
-app.listen(process.env.PORT, () => {
-  console.log(`Server started on port ${process.env.PORT}`);
+dataSource.initialize().then(() => {
+  app.listen(process.env.PORT, () => {
+    console.log(`Server started on port ${process.env.PORT}`);
+  });
 });
