+++ src/shared/typeorm/index.ts
@@ -1,0 +1,13 @@
+import 'dotenv/config';
+import { DataSource } from 'typeorm';
+
+export const dataSource = new DataSource({
+  type: 'postgres',
+  host: 'localhost',
+  port: 5432,
+  username: 'postgres',
+  password: process.env.PASSWORD,
+  database: 'user_management',
+  entities: [],
+  migrations: [],
+});
